Respond to PR comments: new test case for backwards compatibility tests, move timestamp format to types
