+++ src/__test__/deserialization.test.ts
@@ -1,7 +1,7 @@
 import * as fs from "fs";
 import * as path from "path";
 import { parseByteModel, parseFeatherQueryResponse } from "../_bulk_response";
-import { TimestampFormat } from "../_types";
+import { TimestampFormat } from "../_interface";
 
 describe("parseByteModel", () => {
   it("should produce the expected keys", () => {
@@ -46,6 +46,15 @@ describe("parseFeatherQueryResponse", () => {
   it("should handle a multi-query feather response", () => {
     const bytes = _parseByteDataToJSON("binaries/uncompressed_multi.bytes");
     const json = _readJson("json/uncompressed_multi.json");
+    expect(bytes).toMatchObject(json);
+  });
+
+  it("should handle a feather response with epoch_millis timestamp option passed in", () => {
+    const bytes = _parseByteDataToJSON(
+      "binaries/uncompressed_multi.bytes",
+      TimestampFormat.EPOCH_MILLIS
+    );
+    const json = _readJson("json/uncompressed_multi_epoch_millis.json");
     expect(bytes).toMatchObject(json);
   });
